Merge remote-tracking branch 'upstream/main'
So uh... the code went flying everywhere...
Update fork to v0.3.3 resulted in some broken code, specifically in
`SensorManager.cpp`. Lines 60 and 107 reflect this change because the
scan for I2C devices stops the setup. Relevant sections needed for I2C
but not for SPI have been fixed with if statements.